Create TopView of a Binary Tree.cpp
create a class to hold horizontal distance, level and the node. do preorder traversal to assign horizontal distance, level to each node.
sort the vector of custom class that is holding the custom class objects based on horizontal distance, if equal then sort based on levels.
take the first node from the sorted nodes for each horizontal distance, and put the node's value into an answer vector.
that is the answer.