Create construct binary tree using inorder and preorder.cpp
an unordered_map holds the index of each element in the in-order array.
pleft and pright are used to store left and right boundaries of the preorder array.
ileft and iright are used to store left and right boundaries of the inorder array.
'i' holds the current root node's index.
First root is created and assigned the value of pleft of preorder array. Then we calculate no. of elements in left and right sides of the pleft's value in the inorder array.
recursively find for both the left and right subtree.
for left subtree, new pleft is the current pleft + 1. And the new pright is the current pleft + no. of nodes in left side of the inorder array.
for left subtree, new ileft is the current ileft. And the new iright is the (i-1) th index.
for right subtree, new pleft is the current pleft + no. of nodes in left subtree + 1. And the new pright is the current pright.
for right subtree, new ileft is the (i+1)th index. And the new iright is the current 'i' + no. of nodes in the right subtree.
after the recursion is done, the tree is constructed!