Create size if largest bst in a binary tree.cpp
Node is a custom data structure to store minimum value, maximum value, and the max size at that node, including both subtrees.
Now we do post-order traversal, and when we reach the leftmost node, we start by initializing the max, min, and size variables for that node.
Initially, for the leaf nodes, max, and min will be the value of the nodes and the size will be 1.
When we move upwards along the tree, for each node, we check whether the node's value lies between the left node's maximum and the right node's minimum.
Otherwise, that tree can be discarded as not a bst, but we still need to save the max size until that node. (in case there are no larger bst ahead then this itself will be our final answer).
the node's value is valid, then we need to set the node's max as the max of the right subtree, and min as the min of the left subtree.
(here we need to check whether they're not INT_MAX and INT_MIN, otherwise, we'll get some wrong answers).
after doing postorder traversal, we'll have our final answer!