Create topological sort to check for DAG.cpp
do a topological sort using Kahn's algorithm but keep a count of the no. of times the while loop is executed.
if the no. of times the while loop is executed is equal to the no. of vertices, means the while loop is executed only once for each vertex hence the graph is a DAG.
otherwise, the count is 0 since the in-degree of a directed cyclic graph is never 0.