Create no. of islands using dfs.cpp
if we treat all the digits in the grid as a node and adjacent nodes as connected, this question is converted into finding the no. of connected components.
maintain a visited matrix for the dfs traversal, and traverse only if the node is 1.
in the dfs traversal, check for all 4 directions for unvisited nodes if they are 1, and do the dfs call for each.
the direction array holds the directions, which can be used in a for-loop,
dir[i] represents the y-axis and dir[i+1] represents the x-axis for the next direction.
do take care of boundary checks for index out-of-bounds.