Create dp5 ninja training topdown.cpp
there are 3 activities such as 0,1,2.
and the days are indexed as 0,1,2,...,n.
for the nth day, there's no information about the previous day's task hence we take it as 3 (i.e none of the activities were performed)
so, for the nth day, we can choose 0, 1, or 2nd task.
if we choose the ith task for the nth day, we add points[n][i] to the total points and recursively solve for the n-1th day.
Here since we already took the ith task, we cannot do that task for the n-1th day hence will have to choose any other task.
therefore recurrence relation becomes like this --> f(n,prev) = points[n][i] + f(n-1,i)
when n = 0, we have to choose the task that was not done on the previous day, and that which maximizes the points. ( task other than the previous day's which has max points ).