Create dp5 ninja training bottomup.cpp
dp[0][i] represents, choosing the 0th day tasks when we are going to select ith task the next day. (meaning we can't choose ith task on the 0th day).
hence dp[0][i] will be the max of points[0][j] where j != i.
dp[1][0] represents choosing 1st day task, but we cannot select the 0th task since the 0th task is chosen for the next day.
so for dp[1][0], we choose the max of 1st and 2nd tasks which can be calculated by max( dp[0][1] + points[1][1] and dp[0][2] + points[1][2] ).
Where dp[0][1] represents the highest points earned on day 0, if we choose task 1 on day 1.
dp[0][2] represents the highest points earned on day 0, if we choose task 2 on day 1.
To get the final answer, find the max of points earned on nth day, that is max ( dp[n][0], dp[n][1], dp[n][2] ).
for a better intuitive bottom-up answer, refer to DP 7 of Striver on YouTube.