Create dp13 partition equal subset sum.cpp
sum up all the elements, if it is odd, return false because the array can't be divided into 2 subsets with equal sum.
if it is even then divide the sum by 2, that will be the target to find in the given array.
If we find the target that means, we have the elements which sum up to half of the array sum, and the remaining elements will definitely sum up to the other half of the sum and they are the second subset. hence the array is successfully divided into 2 subsets with equal sum.
use the dp12 problem technique to find whether the array elements will sum up to half (target) the sum of the total array elements.