Create dp18 coin change.cpp
For the TopDown Approach,
the base case is when amount = 0, return 0.
when n = 0, recursively solve by picking coins[0] if coins[0] <= amount
else return 1e9 ( since it is a minimization problem).
the recurrence relation is,
helper(n,amount) = min( helper(n-1,amount), 1 + helper(n, amount - coins[n]) ).
Either we don't pick the coin, or we pick the coin.
But even after we pick the coin, there are unlimited same coins are available.

For the Bottom-Up Approach,
if the amount is 0 then no coins are needed, hence dp[i][0] = 0.
To fill the first row of the dp array,
Since 1st row indicates the 0th coin in the coins array, we divide all possible amounts from 1 to amount by the coins[0] (to get no. of coins needed to make up the amount ) if it is divisible.
if it is not divisible, we initialize it with 1e9 for the minimization condition.
We then build up the array for i = 1 to n and for j = 1 to amount, using the formula,
dp[i][j] = min ( dp[i-1][j], 1 + dp[i][j - coins[i]] ).

For the Space Optimization Approach,
We can observe that, at a time, we only need the current row and the previous row, and don't need the whole dp array.