Create dp19 coin change 2.cpp
For the TopDown Approach,
the base case is if amount = 0, then return 1, since we get to reach the amount with the given coins.
if n = 0, then if the amount is divisible by the coins[0] then return 1, because it means we can reach the amount eventually by recursively using the coins[0].
otherwise, the amount is unreachable hence return 0.
the recurrence relation is,
helper(n,amount) = helper(n-1, amount) + helper(n,amount - coins[n]) .

For the BottomUp Approach,
initialize dp[0][0] as 1, and also initialize the first column with 1s. since if the amount is 0, it means we can reach it from whatever combination we've arrived at amount = 0.
to initialize the first row, for i = 1 to amount, if i is divisible by coins[0] then dp[0][i] = 1, otherwise 0.
Now build up the table using the formula,
dp[i][j] = dp[i-1][j] + dp[i][j-coins[i]] .

We can observe that we only use 2 rows hence we can optimize the space to O(amount).