Create dp28 distinct subsequences.cpp
For the Top-Down Approach,
the base case for the recurrence relation is,
if m < 0 then, the string s2 has matched all, hence return 1.
if n < 0 then, the string s1 has been exhausted and since this comes after the above statement, means string s2 is not matched yet, hence return 0.
there are 2 cases for the recurrence relation. 
if nth character matches with the mth character, then we either, 
-->continue to solve recursively in 1...n-1 and 1...m-1. meaning we are considering the nth character as matched and considered for the current subsequence.
or
-->continue to solve recursively in 1...n-1 and 1...m. meaning we are not considering the nth character for the current subsequence.
if the nth character does not match with the mth character.
here we only have one choice which is not considering the nth character as a part of the current subsequence since it doesn't match, and continue to solve recursively for 1...n-1 and 1...m.
the recurrence relation is,
if s1[n] == s2[m] then helper(n,m) = helper(n-1,m-1) + helper(n-1,m).
else helper(n,m) = helper(n-1,m).

For the Bottom-Up Approach,
initialize the first column as 1 and build up the table from 1 to n and 1 to m.
convert the recurrence relation into the dp formula.

For the space optimization, we can see that we only need 2 rows at a time hence we can reduce 2D dp table into 2 -1D vectors.