Create dp29 edit distance.cpp
For the TopDown Approach,
here we have 3 different possible choices, insertion, deletion, or replacement.
so the base case is,
if n < 0 then it means we exhausted the s1 string, hence to make it reach string s2, we need to add the remaining no. of characters in string s2, hence return m+1
if m < 0, it means we reached the end of string s2, hence to make s1 == s2, we need to delete the remaining characters in s1, hence return n+1.

if s1[n] == s2[m] then no need to do anything and we can recursively solve for the s1 from 1...n-1 and s2 from 1...m-1.
else,
we either 
-->insert a character at the end of s1, making the size n+1, but it gets matched with s2[m] and we recursively solve for s1[1...n] and s2[1...m-1].
or
-->delete the s1[n] character and continue solving recursively for s1[1...n-1] and s2[1...m].
or
-->replace the s1[n] character by s2[m] character then we can match them and continue recursively solving for s1[1...n-1] and s2[1...m-1].
for all the above 3 operations, we add 1 to them because they cost 1 operation and we take the minimum of the 3 as the required answer.
recurrence relation is,
if s1[n] == s2[m] then helper(n,m) = helper(n-1,m-1)
else helper(n,m) = 1 + min ( helper(n,m-1), helper(n-1,m), helper(n-1,m-1) ).

For the BottomUp Approach,
the 1st row represents if s1 reaches the end and 1st column represents if s2 reaches the end.
initialize 1st row as dp[0][j] = j, because if s1 reaches the end then no. of operations required is the no. of characters remaining in s2(to add them).
initialize 1st column as dp[i][0] = i, because if s2 reaches the end then no. of operations required is the no. of characters remaining in s1(to delete them).
Now build up the dp table from i = 1 to n and for j = 1 to m using the formula converted from the recurrence relation.

For the Space optimization, we can observe that we only need 2 rows at a time hence we convert the 2D dp table into 2 - 1D vectors.