Create dp30 wildcard matching.cpp
For the TopDown Approach,
the base case is,
if n < 0 and m < 0, means both the strings are exhausted and hence return true. ( s2 is exhausted means there is a match)
if m < 0 and n >= 0, means string s1 is not exhausted, which means there's no match, hence return false.
if n < 0 and m >= 0, means string s2 is not exhausted, hence we need to traverse through the string and check if any of the characters is not *.
if there is a character other than *, then return false, else return true.
There are 2 possibilities in paths,
if s1[n] == s2[m] or s2[m] == '?', then we consider the characters of s1 and s2 as a match and we recursively solve for s1[1...n-1] and s2[1...m-1].
else if s2[m] == ' * ', then there are 2 choices we can make.
--> skip the '*' and don't match any character from s1 to '*' and recursively solve for s1[1...n] and s2[1...m-1].
-->match the character in s1 with the '*' but don't skip the '*' character, that is, recursively solve for s1[1...n-1] and s2[1...m]. (reuse the '*' again).
the recurrence relation is,
if s1[n] == s2[m] then helper(n,m) = helper(n-1,m-1)
else if s2[m] == '*' then helper(n,m) = helper(n,m-1) or helper(n-1,m).
else helper(n,m) = false.