Create dp40 longest bitonic subsequence.cpp
this uses a similar approach as dp 41 and dp 42 lectures of striver's dp series.
we hold 2 vectors, one with the longest increasing subsequence at each index i, and another with the longest decreasing subsequence at index i. (i.e longest increasing subsequence from n-1 to 0).
for i = 0 to n-1, we add increasing[i] and decreasing[i] and subtract 1 from the sum (because the ith element is common to both arrays and is counted twice).
we then find the max of the resulting vector and return it as the answer.