Update dp41 no. of longest increasing subsequences.cpp
refer to dp 47 lecture of striver's dp series.

we keep 2 vectors, longest and count, to keep track of the longest subsequence at index i and the count of the longest subsequence at index i.
at first longest and count both are initialized with 1s because at least the longest increasing subsequence will be 1 always and the count of LIS at index i will always be at least 1.
now we use the dp42 lecture technique to find LIS and when we update the value of each i for i = 1 to n and j = 0 to i-1.
if nums[i] > nums[j] and longest[i] < longest[j] + 1 then we can update the longest[i] = longest[j] + 1.
and the count[i] can be updated with count[j].
This is because when we satisfy the above if condition, it means we can add the nums[i] element to the new longest subsequence and the count of the longest subsequence is usually the same as the count of nums[j]. Because we only have 1 subsequence from nums[j] to nums[i] and there might be many subsequences till nums[j]. So our new count[i] will have as many subsequences as there were for count[j].

if nums[i] > nums[j] and longest[i] == longest[j] + 1, then count[i] += count[j].
This means if the nums[i] can be a part of the subsequence with nums[j] but the length is not the longest but equal to the length of a previously obtained longest[i],
then we need to add the count[j] to the count[i]. This is because the no. of total longest increasing subsequences will be the total no. of subsequences at index j + total subsequences obtained till now at index i. 
For example, if there are 2 longest subsequences at index j and 3 longest subsequences at index i, then it means we can obtain 2 + 3 = 5 different combinations of nums from 0...i-1 that are a part of LIS at index i.
To get the final answer, we need to see all the different longest array values that are equal to the max value i.e longest increasing subsequence.
Then we need to add all the counts of those indexes then we get the final answer.
This is because, the longest increasing subsequences might not end at the same index i, instead, there might be multiple LIS that end at different index i.
and count[i] only holds the no. of LIS that end at i, so we need to consider all different possible LIS at different index i(that might have many no. of LIS ending at them), then add all their counts to get the final answer.