Create dp44 burst balloons.cpp
This problem is very similar to matrix chain multiplication.
but if you take a balloon, burst it, and try to divide the array into 2 subproblems, there will be a problem of subproblems depending on each other, and hence we can't go ahead with that approach.
for example, --> a b c d e --> burst c --> a b d e --> now try to divide the array into a b and d e.
we see that b needs the value of d and vice versa to get the no. of coins.
so we take a reverse approach and choose the last balloon to burst, and since that balloon will stay until the end, it will become the right or left balloon for other balloons to burst.
and when we burst that balloon at the end, we will only need to multiply it by 1 to the left and 1 to the right.
so at each level of bursting the balloons, we can divide the array into 2 subproblems without them being dependent on each other, because they will only be dependent on the next balloon to burst, which we have already solved.
for example, a b c d e --> last to burst is c --> last but one to burst can be either a or b to the left, or d or e to the right.
and if we choose a, then we have its left and right balloon values, left is the 1 which is the boundary, and right is the next balloon to burst which is c. (balloon b is already burst since we are going from bottom to the top in the recursion tree)
 so we get the cost as 1 * a * c.
if we take b as the last but one balloon to burst, then again we have its left and right balloon values, left is the boundary value which is 1, and the right is the balloon that is going to burst next, that is c.
so we get the cost as 1 * b * c.
so whenever we choose the last balloon to burst, we can divide the problem into 2 subproblems that can be solved recursively.
Now we can choose which balloon to burst at the end in N ways if there are N balloons.
the recurrence relation is,
coins = nums[i-1] * nums[k] * nums[j+1].
helper(i,j) = max ( coins + helper(i,k-1) + helper(k+1,j) ).
where k denotes the currently chosen, last-to-burst balloon.
i represents the start and j represents the end of the range of currently accessible balloons to burst
we append 1 at the start and at the end to avoid out-of-bounds cases.