Create bt3 combination sum 3.cpp
Here we can only take k elements together to form a sum n.
and the numbers possible to choose are from 1 to 9.
so we can only add k elements to the answer if their sum gives n. else we don't add anything and backtrack.
we maintain a num variable which makes sure we don't select a number twice, since num variable only goes from 1 to 9 and doesn't go back from the previously picked element.