Create bt8 m coloring problem.cpp
start with node 0, and try all possible colors, if node 0 can be colored with any 1 color, then recurse for node 1 and try all possible colors... continue this for all the nodes.
If we are able to color all the nodes, we will reach ind == n. Then we can return true.
if we tried all possible colors for a node, but still can't color, means it is not possible to color, hence, return false.