Create arr1 count of subarray sum equal to k.cpp
The logic used here is like this,
first, we calculate the prefix sum of each element of the array.
if the (prefix sum - target) at any point, is already present in the map, that means there exists a subarray with sum = target, from that (prefix sum - target) th index to the current index. Hence we can increase the count by 1.
But, since the negative elements are allowed, there is a chance we might get some prefix sum repeatedly in the array, hence instead of just adding 1 to the count, we add the frequency of the (prefix sum - target) th value to the count.
as for the reason to initialize map[0] = 1,
consider the case --> [1,1,1] k=2
prefix sum array is [1,2,3] and at each point (prefix sum - target) value is --> [1-2, 2-2, 3-2]
if we don't initialize map[0] = 1, then (2-2) = 0 will add 0 to the count and we will get a count of one less than the existing subarrays.